CA-118 split parsers
